sara_ai/requirements.txt
Flask==1.1.2
openai==0.27.0
google-api-python-client==2.26.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-auth==2.3.0
microsoftgraph-python==1.0.0
python-zoom-api==0.1.3
oauthlib==3.1.1
requests-oauthlib==1.3.0
requests==2.26.0
python-dateutil==2.8.2
pytz==2021.3
